port into run_genneration
